requirements.txt
mysql-connector-python==9.1.0
PyQt5==5.15.11
PyQt5-Qt5
PyQt5_sip==12.15.0
Unidecode==1.3.8
